docs(compute-samples): added parallel testing support (#7071)
* docs(compute-samples): added parallel testing support

* chore: try Timeout Rule

* chore: Make it a ClassRule

* chore(debug): change to 1 s timeout

* modified cleanup function to include filter

* reverted back to jupiter class timeout rule

* split huge test file to smaller ones

* lint fix and refactored acc to review comments

* force adding junit4 to be used in surefire plugin

* revert junit dependency and adding skiptest

* added skiptest for failsafe plugin
