+++ compute/cloud-client/src/main/java/compute/CreateTemplateWithSubnet.java
@@ -29,11 +29,13 @@ import com.google.cloud.compute.v1.NetworkInterface;
 import com.google.cloud.compute.v1.Operation;
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 public class CreateTemplateWithSubnet {
 
   public static void main(String[] args)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     /*
     TODO(developer): Replace these variables before running the sample.
     projectId: project ID or project number of the Cloud project you use.
@@ -54,7 +56,7 @@ public class CreateTemplateWithSubnet {
   // Create an instance template that uses a provided subnet.
   public static void createTemplateWithSubnet(String projectId, String network, String subnetwork,
       String templateName)
-      throws IOException, ExecutionException, InterruptedException {
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
     try (InstanceTemplatesClient instanceTemplatesClient = InstanceTemplatesClient.create();
         GlobalOperationsClient globalOperationsClient = GlobalOperationsClient.create()) {
 
@@ -87,7 +89,7 @@ public class CreateTemplateWithSubnet {
           .build();
 
       Operation operation = instanceTemplatesClient.insertCallable()
-          .futureCall(insertInstanceTemplateRequest).get();
+          .futureCall(insertInstanceTemplateRequest).get(3, TimeUnit.MINUTES);
 
       Operation response = globalOperationsClient.wait(projectId, operation.getName());
